Integrate friends & invitation features with database
This commit integrates the 'friends' and 'invitation' features into the database. Previously, these features were commented out in the DI.kt, Tables.kt, and other relevant files. New files like FriendsStorage.kt and InvitationsStorage.kt are added to handle these features.

Furthermore, the dependencies for these features are added in the build.gradle.kts file and unnecessary urgencies present in the url for Database.connect method in DI.kt is removed.

Also, the database url is kept persistent during the test cases, it's handled in the build.gradle.kts of the 'application' module, and wdater version has updated to 0.0.4.

All of these changes are made to make the application more feature-rich, modular and logically consistent. They will help in extending these features in the future.